services/web/app/requirements.txt
Flask==1.1.4
flask-mongoengine==1.0.0
Flask-RESTful==0.3.8
autopep8==1.5.7